pyproject.toml
[project]
name = "edinet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.7",
    "chardet>=5.2.0",
    "googletrans>=4.0.2",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "pandas>=2.2.3",
    "pydantic-settings>=2.9.1",
    "stamina>=25.1.0",
]

[dependency-groups]
dev = ["black>=25.1.0", "ruff>=0.12.0"]


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.ruff]
# Run in check-only mode (no auto-fixing)
fix = true
# Set line length to match Black
line-length = 100

# Include .py and .pyi files, as in Black
include = ["*.py", "*.pyi"]

# Exclude the same directories and files as in Black
exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
    "foo.py",
    "**/*example.py"
]

# Select rules for linting
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "ANN", # flake8-annotations for type annotation checks
    "UP",    # pyupgrade (модернизация кода)
    "PL",
    "RUF",
]

lint.ignore = [
    "ANN001", # Ignore missing type annotation for 'self'
    "ANN002", # Ignore missing type annotation for 'cls'
    "ANN204", # Ignore missing type annotation for '__init__'
    "UP031",  # Ignore use of deprecated module
    "UP046",  # Ignore TypeVar usage
]


[tool.ruff.lint.isort]
known-first-party = []
force-sort-within-sections = true
lines-after-imports = 2
combine-as-imports = true
split-on-trailing-comma = true  # include_trailing_comma
force-wrap-aliases = false      #  force_grid_wrap = 0

